requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
graphviz==0.16
gym==0.18.0
gym-retro==0.8.0
httplib2==0.19.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
neat-python==0.92
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
Pillow==7.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.1
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
